chore: receiving circuit input directly
