+++ lgn-messages/src/types/v1/query/mod.rs
@@ -3,6 +3,8 @@ use std::collections::HashMap;
 use alloy_primitives::U256;
 use serde_derive::Deserialize;
 use serde_derive::Serialize;
+use verifiable_db::api::QueryCircuitInput;
+use verifiable_db::api::QueryParameters;
 use verifiable_db::query::computational_hash_ids::PlaceholderIdentifier;
 use verifiable_db::query::universal_circuit::universal_circuit_inputs::Placeholders;
 
@@ -17,6 +19,42 @@ pub const NUM_CHUNKS: usize = 66;
 /// Maximum number of rows that can be proven in a single proof of batching query
 /// We must use the same value of this constant for both DQ and Worker.
 pub const NUM_ROWS: usize = 100;
+
+pub const ROW_TREE_MAX_DEPTH: usize = 25;
+pub const INDEX_TREE_MAX_DEPTH: usize = 26;
+pub const MAX_NUM_RESULT_OPS: usize = 20;
+pub const MAX_NUM_RESULTS: usize = 10;
+pub const MAX_NUM_OUTPUTS: usize = 5;
+pub const MAX_NUM_ITEMS_PER_OUTPUT: usize = 5;
+pub const MAX_NUM_PLACEHOLDERS: usize = 5;
+pub const MAX_NUM_COLUMNS: usize = 20;
+pub const MAX_NUM_PREDICATE_OPS: usize = 20;
+
+pub type ConcreteCircuitInput = QueryCircuitInput<
+    NUM_CHUNKS,
+    NUM_ROWS,
+    ROW_TREE_MAX_DEPTH,
+    INDEX_TREE_MAX_DEPTH,
+    MAX_NUM_COLUMNS,
+    MAX_NUM_PREDICATE_OPS,
+    MAX_NUM_PREDICATE_OPS,
+    MAX_NUM_OUTPUTS,
+    MAX_NUM_ITEMS_PER_OUTPUT,
+    MAX_NUM_PLACEHOLDERS,
+>;
+
+pub type ConcreteQueryParameters = QueryParameters<
+    NUM_CHUNKS,
+    NUM_ROWS,
+    ROW_TREE_MAX_DEPTH,
+    INDEX_TREE_MAX_DEPTH,
+    MAX_NUM_COLUMNS,
+    MAX_NUM_PREDICATE_OPS,
+    MAX_NUM_RESULT_OPS,
+    MAX_NUM_OUTPUTS,
+    MAX_NUM_ITEMS_PER_OUTPUT,
+    MAX_NUM_PLACEHOLDERS,
+>;
 
 #[derive(Deserialize, Serialize)]
 #[serde(tag = "type")]
